dashboard/requirements.txt
Flask==2.3.2
python-dateutil==2.8.2
